Revert "Typescript attempt #1 (#51)"
This reverts commit 761f8b0b0cc41caaa45c5dd25437c3fea8a4de6e.
