+++ .gitignore
@@ -11,6 +11,3 @@ client/node_modules
 client/DS_Store
 /client/.env
 populate.js
-./dist
-/dist
-dist
