+++ client/package.json
@@ -57,9 +57,8 @@
       "last 1 safari version"
     ]
   },
-  "proxy": "http://localhost:5001",
+  "proxy": "http://localhost:5000",
   "devDependencies": {
-    "@types/styled-components": "^5.1.25",
-    "@babel/plugin-proposal-private-property-in-object": "^7.16.7"
+    "@types/styled-components": "^5.1.25"
   }
 }
