+++ package.json
@@ -2,21 +2,24 @@
   "name": "myanimecollection",
   "version": "1.0.1",
   "description": "A anime tracking application",
-  "main": "server.js",
+  "main": "index.js",
   "type": "module",
   "scripts": {
     "install-server": "npm install",
-    "install-dependencies": "npm run install-client && npm run install-server",
+    "install-dependencies": " npm run install-client && npm run install-server ",
     "build-client": "cd client && npm run build",
     "install-client": "cd client && npm install",
-    "heroku-postbuild": "npm run install-client && npm run build-client && npx tsc",
-    "start-server": "node ./dist/server.js",
-    "start-client": "npm start --prefix client",
-    "start": "concurrently --kill-others-on-fail \"npm run start-server\" \"npm run start-client\""
+    "heroku-postbuild": "npm run install-client && npm run build-client",
+    "server": "nodemon server --ignore client",
+    "client": "npm start --prefix client",
+    "start": "concurrently --kill-others-on-fail \" npm run server\" \" npm run client\""
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
+  "devDependencies": {
+    "nodemon": "^2.0.20"
+  },
   "dependencies": {
     "@emotion/react": "^11.10.0",
     "@emotion/styled": "^11.10.0",
@@ -39,7 +42,6 @@
     "moment": "^2.29.4",
     "mongoose": "^6.12.0",
     "morgan": "^1.10.0",
-    "nodemon": "^2.0.22",
     "normalize.css": "^8.0.1",
     "react": "^18.2.0",
     "react-dom": "^18.0.0",
@@ -56,10 +58,5 @@
   },
   "engines": {
     "node": "16.13.0"
-  },
-  "devDependencies": {
-    "@types/dotenv": "^8.2.0",
-    "@types/helmet": "^4.0.0",
-    "@types/morgan": "^1.9.5"
   }
 }
