--- tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "compilerOptions": {
-    "outDir": "./dist",
-    "rootDir": "./",
-    "target": "ES2020",
-    "module": "Node16",
-    "moduleResolution": "Node16",
-    "baseUrl": "./",
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
-    "strict": true,
-    "skipLibCheck": true,
-    "allowJs": true
-  },
-  "include": ["./**/*.ts"],
-  "exclude": ["node_modules", "dist", "client"]
-}
